requirements.txt
anthropic>=0.45.0
httpx>=0.27.0
opik>=1.0.0
beautifulsoup4>=4.12.0
jsonschema>=4.20.0
python-dotenv>=1.0.0
feedparser>=6.0.0
